pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lol-langgraph-agent"
version = "0.1.0"
description = "Agentic LoL Assistant using LangGraph, Gemini Flash, and MCP."
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "langchain>=0.1.13",
    "langgraph>=0.0.30",
    "langchain-google-genai>=1.0.1",
    "langchain-community>=0.0.29",
    "langchain-tavily>=0.1.0",
    "langchain-mcp-adapters>=0.2.2",
    "pydantic>=2.6.4",
    "python-dotenv>=1.0.1",
    "langchain-ollama>=0.2.0"
]

[tool.hatch.build.targets.wheel]
packages = ["app"]